chore: remove Details as composable; expose item and weekdays
