+++ lib/components/Widgets/exports.tsx
@@ -1,9 +1,10 @@
 export * from "./CategoryBarSection"
 export * from "./Charts/exports"
 export * from "./Dashboard"
-export * from "./Details"
+export * from "./DetailsItem"
 export * from "./IndicatorsList"
 export * from "./Info"
 export * from "./TimeStatus"
+export * from "./Weekdays"
 export * from "./WidgetContainer"
 export * from "./WidgetStrip"
